apps/agent-os/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
psycopg2-binary==2.9.9
openai==1.10.0
composio==0.8.0
websockets==12.0
python-dotenv==1.0.0
python-multipart==0.0.6
requests==2.31.0
